% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prikage-package.R
\docType{package}
\name{prikage-package}
\alias{prikage}
\alias{prikage-package}
\title{prikage: prikage(Priya + Package), Embeds my Shiny Application & Launches its via a Function}
\description{
This package prikage, is my first ever R package this packages allows the to launch my first ever shiny application through the launch_app function exported by my package, prikage. The required shiny application has already been embedded my package.This package, gives the user and overview of COVID-19, by the exciting graphs embedded in this application and added user interactivity. Download this package and know more about its detailed functionality.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/etc5523-2020/r-package-assessment-priya51096.git}
  \item Report bugs at \url{https://github.com/etc5523-2020/r-package-assessment-priya51096/issues}
}

}
\author{
\strong{Maintainer}: Priya Ravindra Dingorkar \email{pdin0008@student.monash.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
