% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_functions.R
\name{select_input_austate}
\alias{select_input_austate}
\title{This is on eof the funct for my app}
\usage{
select_input_austate(B, aus_state)
}
\description{
This is on eof the funct for my app
}
\details{
ONe of the select input trajectory
}
