% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_functions.R
\name{date_slider_input}
\alias{date_slider_input}
\title{This is on eof the funct for my app}
\usage{
date_slider_input(C)
}
\description{
This is on eof the funct for my app
}
\details{
ONe of the select input trajectory
}
