% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_gg_functions.R
\name{gg_covid_barplot}
\alias{gg_covid_barplot}
\title{gg_covid_barplot functions creates a bar chart that shows us the covid- 19 positive case for the user selected state of Australia}
\usage{
gg_covid_barplot(aus_state, input)
}
\arguments{
\item{aus_state}{is the required argument of this function, to help the function recognize the data and further work on it as per the function}

\item{input}{allows the user to read the data for filtering/matching as per the function in order to produce the graph}
}
\description{
gg_covid_barplot function helps us for the creation of bar charts of the covid confirmed cases for one of the Australia's selected state
ggplot2 and geom_line, are the dependencies of this functions, we has also set a theme for the graphs which allows the  uniformity, throughout our app.
We further add this function in the ggplotly function to enhances the user interactivity of our graph.'
}
