% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_gg_functions.R
\name{gg_covid_line}
\alias{gg_covid_line}
\title{gg function}
\usage{
gg_covid_line(covid_data, input)
}
\description{
gg function for the line graphs of the covid confirmed, recovered and death cases
}
