% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_functions.R
\name{select_input_austate}
\alias{select_input_austate}
\title{select_input_austate allows the user to select one of the states of Australia.}
\usage{
select_input_austate(selected_state_label, aus_state)
}
\arguments{
\item{selected_state_label}{defines a new label everytime this function runs and save it's value in the input_id of our select input function}

\item{aus_data, }{this requires the function to pass the aus_data, which will help the function recognize other select input default parameter, for the choices}
}
\description{
select_input_austate function allows the user to choose from a list of all the Australia's state
The data used for this function is a external data for a github repository. (Available in the data folder of this application)
select_input_ausstate function, accepts and stores the state of Australia as per user's choice.
The user's selected state can be stored in as a label in the selected_state_label, which is one of the argument of our function
The user also will have to pass aus_state which is already been stored in the app.R for the second argument of this function
}
